장소 제공 구현 + LLM 추천 이유
