--- src/main/java/com/itplace/userapi/ai/config/AIConfig.java
@@ -1,21 +1,0 @@
-package com.itplace.userapi.ai.config;
-
-import org.springframework.ai.chat.memory.ChatMemoryRepository;
-import org.springframework.ai.chat.memory.repository.jdbc.JdbcChatMemoryRepository;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.transaction.PlatformTransactionManager;
-
-@Configuration
-public class AIConfig {
-
-    @Bean
-    public ChatMemoryRepository chatMemoryRepository(JdbcTemplate jdbcTemplate, PlatformTransactionManager transactionManager) {
-        return JdbcChatMemoryRepository.builder()
-                .jdbcTemplate(jdbcTemplate)
-                .transactionManager(transactionManager)
-                .build();
-    }
-
-}
