--- src/main/java/com/itplace/userapi/ai/dto/RecommendReason.java
@@ -1,10 +1,0 @@
-package com.itplace.userapi.ai.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class RecommendReason {
-    private String reason;
-}
