--- src/main/java/com/itplace/userapi/ai/entity/ChatHistory.java
@@ -1,39 +1,0 @@
-package com.itplace.userapi.ai.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import lombok.Getter;
-import lombok.Setter;
-import org.hibernate.annotations.CreationTimestamp;
-import org.springframework.ai.chat.messages.MessageType;
-
-@Entity
-@Getter
-@Setter
-@Table(name = "chatHistory")
-public class ChatHistory {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "userId")
-    private String userId;
-
-    @Column(name = "content", columnDefinition = "TEXT")
-    private String content;
-
-    @Enumerated(EnumType.STRING)
-    private MessageType type;
-
-    @CreationTimestamp
-    @Column(name = "createdAt")
-    private LocalDateTime createdAt;
-}
