+++ src/main/java/com/itplace/userapi/ai/llm/dto/RecommendReason.java
@@ -1,0 +1,10 @@
+package com.itplace.userapi.ai.llm.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class RecommendReason {
+    private String reason;
+}
