+++ src/main/java/com/itplace/userapi/ai/question/index/QuestionIndexer.java
