+++ src/main/java/com/itplace/userapi/ai/rag/config/ElasticsearchConfig.java
@@ -1,0 +1,36 @@
+package com.itplace.userapi.ai.rag.config;
+
+import co.elastic.clients.elasticsearch.ElasticsearchClient;
+import co.elastic.clients.json.jackson.JacksonJsonpMapper;
+import co.elastic.clients.transport.ElasticsearchTransport;
+import co.elastic.clients.transport.rest_client.RestClientTransport;
+import org.apache.http.HttpHost;
+import org.elasticsearch.client.RestClient;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ElasticsearchConfig {
+
+    @Value("${elasticsearch.host}")
+    private String host;
+
+    @Value("${elasticsearch.port}")
+    private Integer port;
+
+    @Value("${elasticsearch.scheme}")
+    private String scheme;
+
+    @Bean
+    public ElasticsearchClient elasticsearchClient() {
+
+        RestClient restClient = RestClient.builder(new HttpHost(host, port, scheme)).build();
+
+        ElasticsearchTransport transport = new RestClientTransport(
+                restClient, new JacksonJsonpMapper());
+
+        return new ElasticsearchClient(transport);
+    }
+}
+
