+++ src/main/java/com/itplace/userapi/ai/rag/index/ElasticsearchIndexer.java
