+++ src/main/java/com/itplace/userapi/ai/rag/service/ElasticService.java
@@ -1,0 +1,7 @@
+package com.itplace.userapi.ai.rag.service;
+
+public interface ElasticService {
+    void createIndexIfNotExists(String indexName);
+
+    boolean existsById(String indexName, String id);
+}
