+++ src/main/java/com/itplace/userapi/ai/rag/service/EmbeddingService.java
@@ -1,0 +1,7 @@
+package com.itplace.userapi.ai.rag.service;
+
+import java.util.List;
+
+public interface EmbeddingService {
+    List<Float> embed(String text);
+}
