--- src/main/java/com/itplace/userapi/ai/repository/ChatHistoryRepository.java
@@ -1,9 +1,0 @@
-package com.itplace.userapi.ai.repository;
-
-import com.itplace.userapi.ai.entity.ChatHistory;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ChatHistoryRepository extends JpaRepository<ChatHistory, Long> {
-    List<ChatHistory> findByUserIdOrderByCreatedAtAsc(String userId);
-}
