--- src/main/java/com/itplace/userapi/question/service/ElasticQuestionService.java
@@ -1,9 +1,0 @@
-package com.itplace.userapi.question.service;
-
-public interface ElasticQuestionService {
-    void createQuestionIndexIfNotExists(String indexName);
-
-    boolean existsById(String indexName, String id);
-
-    void saveQuestion(String indexName, String id, String question, String category, java.util.List<Float> embedding);
-}
