--- src/main/java/com/itplace/userapi/rag/service/BenefitSearchService.java
@@ -1,9 +1,0 @@
-package com.itplace.userapi.rag.service;
-
-import com.itplace.userapi.benefit.entity.enums.Grade;
-import com.itplace.userapi.recommend.dto.Candidate;
-import java.util.List;
-
-public interface BenefitSearchService {
-    List<Candidate> queryVector(Grade grade, List<Float> userEmbedding, int topK);
-}
