--- src/main/java/com/itplace/userapi/rag/service/ElasticService.java
@@ -1,7 +1,0 @@
-package com.itplace.userapi.rag.service;
-
-public interface ElasticService {
-    void createIndexIfNotExists(String indexName);
-
-    boolean existsById(String indexName, String id);
-}
