--- src/main/java/com/itplace/userapi/rag/service/EmbeddingService.java
@@ -1,7 +1,0 @@
-package com.itplace.userapi.rag.service;
-
-import java.util.List;
-
-public interface EmbeddingService {
-    List<Float> embed(String text);
-}
